pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sistema-b"
version = "0.1.0"
description = "Control de versiones simple para modelos BP locales"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "gitpython>=3.1.40",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
]

[project.scripts]
bp = "bp.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
bp = ["web/templates/*.html", "web/static/*.css", "web/static/*.js"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
